refactor(studio): split the health panel's digest and share its SLO formatting
HealthPanel was 346 lines, now 282.

- `HealthDigest` — the two-column "what is going wrong right now" grid (functions
  by error rate, most recent errors). Both cards render data the panel already
  fetched and read none of its fan-out, poll, or shard state.
- `slo-format.ts` — the level type, its badge tones, the warn/crit thresholds,
  and the rate renderer, so the panel and the digest cannot disagree on what a
  breach looks like.

Suppression deleted. 79 reports tests green.

Refs #230.
